Create .codeclimate.yml
Hi this is Emily from Code Climate Support. There's more info in the email I've sent about this pull request, but wanted provide this in a PR, too.

Before merging this, install a webhook for Code Climate so it can see subsequent commits:

- you can do with these [instructions](https://docs.codeclimate.com/docs/installing-code-climates-webhook).

Then, merge this inferred .codeclimate.yml with the prepare step that fetches your external Rubocop config settings.